html/cross-origin-embedder-policy/sandbox.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<div id=log></div>
<script>
async_test(t => {
  window.addEventListener("message", t.step_func_done(({ data }) => {
    assert_equals(data.origin, "null");
    assert_true(data.sameOriginWithoutCORP, "Request to same-origin resource without CORP did not fail");
    assert_true(data.sameOriginWithSameOriginCORP, "Request to same-origin resource with same-origin CORP did not fail");
    assert_true(data.sameOriginWithCrossOriginCORP, "Request to same-origin resource with cross-origin CORP did not succeed");
    assert_true(data.crossOriginWithCrossOriginCORP, "Request to cross-origin resource with cross-origin CORP did not succeed");
  }));

  const origins = get_host_info();
  const frame = document.createElement("iframe");
  const nothingPath = new URL("resources/nothing.txt", window.location).pathname;
  const nothingSameOriginPath = new URL("resources/nothing-same-origin-corp.txt", window.location).pathname;
  frame.sandbox = "allow-scripts";
  frame.srcdoc = `<script>
let i = 0;
const data = { sameOriginWithoutCORP: false,
               sameOriginWithSameOriginCORP: false,
               sameOriginWithCrossOriginCORP: false,
               crossOriginWithCrossOriginCORP: false,
               origin: self.origin };
function report(token, success) {
  i++;
  data[token] = success;
  if (i == 4) {
    parent.postMessage(data, "*");
  }
}
fetch("/common/blank.html", { mode: "no-cors" }).then(() => report("sameOriginWithoutCORP", false), () => report("sameOriginWithoutCORP", true));
fetch("${nothingSameOriginPath}", { mode: "no-cors" }).then(() => report("sameOriginWithSameOriginCORP", false), () => report("sameOriginWithSameOriginCORP", true));
fetch("${nothingPath}", { mode: "no-cors" }).then(() => report("sameOriginWithCrossOriginCORP", true), () => report("sameOriginWithCrossOriginCORP", false));
fetch("${origins.HTTPS_NOTSAMESITE_ORIGIN}${nothingPath}", { mode: "no-cors" }).then(() => report("crossOriginWithCrossOriginCORP", true), () => report("crossOriginWithCrossOriginCORP", false));
<\/script>`;
  document.body.append(frame);
}, "Cross-Origin-Embedder-Policy and sandbox");
</script>
